requirements.txt
numpy~=1.26.3
pandas~=2.2.2
matplotlib~=3.9.3
netCDF4~=1.7.2
pillow~=11.0.0
torch~=2.7.0
torchvision~=0.22.0